refactor: split read query package
